add config for library
