+++ src/index.ts
@@ -1,0 +1,3 @@
+export { EmitInterface } from './pages/emitInterface';
+export type { EmitInterfaceProps,EmitInterfaceConfig } from './pages/emitInterface/types';
+
