--- src/lib/index.ts
@@ -1,10 +1,0 @@
-import { EmitInterface } from '../pages/emitInterface';
-
-// Export the main component
-export { EmitInterface };
-
-// Export types
-export type { EmitInterfaceProps } from '../pages/emitInterface/types';
-
-// Export default for easier imports
-export default EmitInterface;
